refactor: extracted test utils and formatting
